Added Xft support (in the same fashion as done in dmenu by Hiltjo Posthuma), Xinerama support, changed colors and key layout
This is version 6 of my svkbd patch, and I'm nearing completion on all
the stuff I wanted to implement. I decided to split this patch into 4
logical units, making for a nicer commit history when merged,
all four together constitute all the work I did on svkbd
and need to be applied in succession.

See the last patch in the series (4/4) for comments on what I changed since
v5.

Signed-off-by: Miles Alan <m@milesalan.com>